feat: java 오류 수정
