--- programmers/난이도별/level02.H-Index/README.md
@@ -1,3 +1,0 @@
-# 출처
-
-https://programmers.co.kr/learn/courses/30/lessons/42747
