+++ programmers/난이도별/level02.기능개발/sjjyy.java
@@ -1,3 +1,5 @@
+package programmers.난이도별.level02.기능개발;
+
 import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.Queue;
