+++ programmers/난이도별/level02/H_Index/rockmiin.py
